// packages/backend/src/database/types/anexo.ts
/**
 * AUTO-GENERATED FILE @ 2020-06-21 04:08:56 - DO NOT EDIT!
 *
 * This file was automatically generated by schemats v.3.0.3
 * $ schemats generate -c postgres://username:password@localhost/reloaded -t anexo -s public
 *
 */

    export type id = number;
    export type created_at = Date | null;
    export type changed_at = Date | null;
    export type extensao = string;
    export type url = string;
    export type tempo_duracao = string | null;
    export type descricao = string | null;
    export type tipo = 'documento' | 'midia' | null;
    export type fk_imovel_id = number | null;
    export type fk_vistoria_id = number | null;

export interface Anexo {
    id: id;
    created_at: created_at;
    changed_at: changed_at;
    extensao: extensao;
    url: url;
    tempo_duracao: tempo_duracao;
    descricao: descricao;
    tipo: tipo;
    fk_imovel_id: fk_imovel_id;
    fk_vistoria_id: fk_vistoria_id;

}